Bug #1600351 coverity fix
[root cause  ] paramter not match
[changes     ] /bsp/modules/camera
[side effects] NA
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums9620
[test case]:camera case
[reviewers   ] eddy.wei
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ interface/sprd_isp_2v6.h
@@ -2787,8 +2787,8 @@ struct debug_base_info {
 	int32_t isp_cid;
 	int32_t scene_id;
 	int32_t frame_id;
-	unsigned long sec;
-	unsigned long usec;
+	int32_t sec;
+	int32_t usec;
 	int32_t size;
 	int32_t res_data0[8];
 	uint32_t awbc_r;
